feat: Added a Interface for scope to have their own way of handling items
This way, each item expressions queries for State for getting the current item. The state find the itemhandler in the internal list, at calls that itemhandler, to return the current item. The itemahandler is meant to be the for each action, which needs to keep track of which item, that is the current item.

Fixes #20 Fixes #22